refactor: parameterize register_decision_variant! size fields
The macro previously hardcoded num_vertices/num_edges as size getters,
which only works for graph problems. Now accepts dims, fields, and
size_getters as parameters so non-graph Decision variants can specify
their own problem-size fields (e.g., num_vars/num_clauses for SAT).

Callers define inherent methods on Decision<P> before invoking the macro.
